From Bertrand: Fix a problem in TGaxis due to a special behaviour of "mktime" on Windows
If successful, *mktime* sets the values of *tm_wday* and *tm_yday* as
appropriate and sets the other components to represent the specified
calendar time, but with their values forced to the normal ranges. The final
value of *tm_mday* is not set until *tm_mon* and *tm_year* are determined.
When specifying a *tm* structure time, set the *tm_isdst* field to:

   * Zero (0) to indicate that standard time is in effect.
   * A value greater than 0 to indicate that daylight savings time is
     in effect.
   * A value less than zero to have the C run-time library code compute
     whether standard time or daylight savings time is in effect.

(The C run-time library assumes the United States' rules for implementing
the calculation of Daylight Saving Time.) *tm_isdst* is a required field. If
not set, its value is undefined and the return value from *mktime* is
unpredictable. If /timeptr/ points to a *tm* structure returned by a
previous call to *asctime*, *gmtime*, or *localtime*, the *tm_isdst* field
contains the correct value.


git-svn-id: https://root.cern.ch/svn/root/trunk@14934 27541ba8-7e3a-0410-8455-c3a389f83636