Add new package tmva
Authors (alphabetical):
      Andreas Hoecker <Andreas.Hocker@cern.ch> - CERN, Switzerland
      Xavier Prudent  <prudent@lapp.in2p3.fr>  - LAPP, France
      Helge Voss      <Helge.Voss@cern.ch>     - MPI-KP Heidelberg, Germany
      Kai Voss        <Kai.Voss@cern.ch>       - U. of Victoria, Canada

The package is fully documented on the TMVA source forge web site at
  http://tmva.sourceforge.net/

see the ROOt html documentation at:
  http://root.cern.ch/root/htmldoc/TMVA_Index.html

The Toolkit for Multivariate Analysis (TMVA) provides a ROOT-integrated
environment for the parallel processing and evaluation of MVA techniques
to discriminate signal from background samples. It presently includes
(ranked by complexity):

          o Rectangular cut optimisation
          o Correlated likelihood estimator (PDE approach)
          o Multi-dimensional likelihood estimator (PDE - range-search approach)
          o Fisher (and Mahalanobis) discriminant
          o H-Matrix (chi-squared) estimator
          o Artificial Neural Network (two different implementations)
          o Boosted Decision Trees

The TMVA package includes an implementation for each of these discrimination
techniques, their training and testing (performance evaluation).
In addition all these methods can be tested in parallel, and hence their
performance on a particular data set may easily be compared. For the
comparison of the efficiency and background rejection of all the different
methods, the analysis job gives some tabulated efficiency / background values
in the log file, as well as the efficiencye vs. background rejection curves
along with detailed other evaluation information (like ranking, correlation
matrixes and alike) in a ROOT file. These results can then be easily displayed
using the ROOT macros provided with TMVA.

The C++ code runs alternatively as a standalone executable, or as a ROOT
script, where libTMVA.so is linked as a shared library. Each method that
is trained in this application writes the training results in individual
"weight" files, which are either text or ROOT files. Two single compact
C++ classes (reader/TMV_Reader and reader/TMVA_ReaderTools) are provided,
which interpret the training files, and which can be included in any C++ or
ROOT-based analysis job.

Special emphasis has been put on the comparative performance assessment of
the various MVA methods. The training, testing and evaluation phases are
performed in parallel for the various methods (through a factory).
The evaluation accomodates several numerical performance estimators as well
as various ROOT plots. ROOT scripts to conveniently access these plots are
provided.

Please report any problems and/or suggestions for improvements to the authors.


git-svn-id: https://root.cern.ch/svn/root/trunk@14948 27541ba8-7e3a-0410-8455-c3a389f83636