From Gerri: TDSet changes to have the name being the name.
The constructor takes an additional argument. The first argument can be
either the real name of the data set (with type specified by the 4th argument -
default "TTree"), or the type, as before (the rule is that if the 1st arg
is the name of a known class and the 4th arg is undefined, the 1st arg is
taken as type; to use the name of a known class as dataset name the real
type of the class must be specified as 4th arg, i.e.

     TDSet("TTree","h41",0,"TChain")

creates a dataset named "TTree" where the objects to be analyzed are named
"h41" and are of type "TChain", while

     TDSet("TTree","h41")

will create a dataset named "h41" where the objects to be analyzed are named
"h41" and are of type "TTree", as before (the concept of 'name' was not
there; the default name for the dataset is the object name).

It is be backward compatible.


git-svn-id: https://root.cern.ch/svn/root/trunk@15290 27541ba8-7e3a-0410-8455-c3a389f83636