From Axel: If THtml knows a class A, but can't find its sources, it assumed that it's a ROOT class, and the user was creating documentation for his/her own classes. This fails if a user dares to have more than one lib. There was a mechanism implemented which would allow for multiple libs, but that was really a hack.
THtml now checks the env var Root.Html.LibName, and creates a link using
that base URL, for all classes being defined in libLibName.*. I.e. if
that class A is defined in libLibName, and Root.Html.LibName is
../LibNameDoc, then THtml will create a link to ../LibNameDoc/A.html.
The class name / library name mapping is done with the help of rlibmap;
see the class doc for more info.

Other changes:
* fix javascript error in source files
* link javascript file from the proper directory
* reduce width of divs to slightly < 100%
* a names (anchors) should be id-compatible, and thus start with a letter
* update doc.


git-svn-id: https://root.cern.ch/svn/root/trunk@16097 27541ba8-7e3a-0410-8455-c3a389f83636