From Andreas: TAlienFile inherits now directly from TXNetFile. This was possible by introducing another factory function in TAlienFile (TAlienFile::Open), moving some private to protected members and changing some function to virtual in the parent classes. For the new factory function of the plubing rootrc.in had to be modified.
Using the occasion, I have also implemented that, that
TAlienFile tries automatically to open any of the existing images of a
file, starting with the 'closest' one, if it fails with the next
replica aso. This is implemented right now only for reading. It is not
ought to do loadbalancing by random shuffling of the SE order.

There are some changes in TAlienCollection for Panos, to build the 'and'
of two TAlienCollections and to read evelist from the xml file.

Some memory leaks have been fixed in TAlienResult/TGridResult etc.

Finally to have a consistent file name (e.g. a TAlienFile has it's LFN as
a filename - eventhough it inherits from TXNetFile, which does not like
LFN's), I had to introduce another argument to TXNetFile (in the end with
a default value), to tell TXNetFile, that is has to open an xrootd url,
but it has to 'carry' another name. The normal functionality is not
touched by this.


git-svn-id: https://root.cern.ch/svn/root/trunk@16438 27541ba8-7e3a-0410-8455-c3a389f83636