From Markus/Pere:
A problem has been detected by Ioannis running the POOL regression tests with
the latest ROOT production release. The problem shows up in a rather complex
model (CMS inspired I think) involving a vector of pairs containing objects
embedding pool::Refs in split mode. The wrong offset is calculated for the
pool::Ref producing a core dump when doing a dynamic_cast.

The origin of the problem is the fix made for a bug detected by CMS when
classes with reflex dictionaries contained embedded objects described only
with CINT dictionaries.

Markus has debugged the problem and the fix is attached to this report. It is
important to produce a patch version of ROOT 5.14 with this fix.


git-svn-id: https://root.cern.ch/svn/root/trunk@17238 27541ba8-7e3a-0410-8455-c3a389f83636