TMath::Abs, TMath::Min, TMath::Max, TMath::Sign, TMath::Range These functions are unfortunately not defined in a standard way in std::
This include is referenced by a new version of TMath.h.
As a result, TMath.h is back compatible with the previous version.

TMathBase.h is used in place of TMath.h in all the classes
that will go into the future miniCore library.
TMath.h and the TMath implementation will go into a new math sub-directory.

TString.h uses TMathBase.h instead of TMath.h.
As a result, it was necessary to include "TMath.h" in some classes
assuming that TMath was included via TString and using other functions
than the ones defined in TMathBase.h
 ----------------------------------------------------------------------


git-svn-id: https://root.cern.ch/svn/root/trunk/proof@17306 27541ba8-7e3a-0410-8455-c3a389f83636