From Olivier and me: Fix in TTimeStamp::GetZoneOffet(). The offset on Mac OS X and Free and Open BSD was the oposite from Linux/Unix and Windows. The offset returned by this routine is now defined as:
This is the value in seconds one must add to the local time to arrive at
Coordinated Universal Time, so it is negative east of the Prime Meridian.

the difference was due in the way the offset was obtained on the different
platforms. The redefinition results in the offset flipping sign on Mac/BSD
to be in lign with the other platforms.


git-svn-id: https://root.cern.ch/svn/root/trunk@17837 27541ba8-7e3a-0410-8455-c3a389f83636