From Valeri Fine:
Allow the symbol "+" to be a part of the file directory
name. There was a clash with the ROOT file name convention
to treat "+" as a flag of the "animated" pixmap file.
-----
PS: STAR NFS directory do have the "+" in the directory
name. This is why the problem was discovered.


git-svn-id: https://root.cern.ch/svn/root/trunk@18149 27541ba8-7e3a-0410-8455-c3a389f83636