From Roel Aaij: New class TGSplitButton - a split button control has a button and a menu. The menu includes predefined commands for secondary functionality.
Consider the following when you include split buttons in an application:
There are 2 modes available that may be changed by "on the fly" by
calling the SetSplit(Bool_t) method.

If the button is split (i.e. SetSplit(kTRUE) was called), a menu will
popup when the menu area of the button is clicked (the right part
containing a small triangle). Activating a menu item changes the
functionality of the button by having it emit a additional signal
when it is clicked. The signal emitted when the button is clicked,
is the ItemClicked(Int_t) signal with a different fixed value for
the Int_t that corresponds to the id of the activated menu entry.
The button label should update to reflect the last item selected
from the menu.

If the button is not split, clicking it will popup the menu and the
ItemClicked(Int_t) signal will be emitted when a menu entry is
acitvated. The value of the Int_t is again equal to the value of
the id of the activated menu entry.

o Added protection for deletion of tooltip TGButton dtor


git-svn-id: https://root.cern.ch/svn/root/trunk@19397 27541ba8-7e3a-0410-8455-c3a389f83636