From Stefan and me: Move to gccxml version 0.7.0 (checkout date 20070615).
The installation of the LCG external package was done with the checkout opt -D20070615. In addition the configuration file for gccxml was patched so that "gccxml --version" will return "0.7.0_20070615" instead of "0.7.0". This version is also checked in genreflex and a warning is printed if only "0.7.0" is received as version info. This will allow a checkout and build of the gccxml package from CVS.

Changes in gendict.py
- "access" is not defaulted to "public" anymore but has to be checked explicitly
- new attribute "demangled" for each type in the xml output
- unnamed types have no name anymore (the internal name is derived from the "demangled" attribute
- the name of the anonymous namespace is now "@anonymous@namespace@" (was the internal name before)

Changes in genreflex.py
- Change gccxml version from "0.6.0_patch3" to "0.7.0_20070615"

!!! these changes are BACKWARDS INCOMPATIBLE with gccxml the 0.6.0 series !!!
(But we provide the old, deprecated 06.py files for now)


git-svn-id: https://root.cern.ch/svn/root/trunk@19509 27541ba8-7e3a-0410-8455-c3a389f83636