From Zoltan.Torzsok@cern.ch: *** Using constants in GDML with ROOT ***
[ORIGINAL APPROACH]

Originally constants could not be used within mathematical expressions in ROOT.
The following example is to illustrate the problem:

<define>
	<constant name="size" value="100.0"/>
</define>

<solids>
	<box name="MyCube" x="size" y="size" z="size"/>
	<box name="MyBox" x="1.0*size" y="2.0*size" z="3.0*size"/>
</solids>

The "MyCube" solid can be parsed since it contains a constant only.
The "MyBox" solid will indicate an error message like "size" is not
recognised as an expression. This happens because in the original method
the whole expression string is searched in the constant map.
If it can be found, it is replaced with its value.
In "MyBox" the expression string for the "x" attribute is "1.0*size"
and there isnt any constant defined with the name of "1.0*size".
We can trick the parser and eliminate the mentioned error message by
adding the following constants:

<constant name="1.0*size" value="100"/>
<constant name="2.0*size" value="200"/>
<constant name="3.0*size" value="300"/>

[MY APPROACH]

I am using a find-and-replace method on the expression string.
If the name of a constant is found within the expression string,
it is replaced with the value of the constant.

This approach fails in special cases, when a constants name contains
an other constants name, like in the following:

<constant name="wi" value="100"/>
<constant name="width" value="200"/>

In this case, every occurence of "width" will look like "100dth"
so that "width" will be no longer recognised. This must be fixed.

[TEST RESULTS]

Applying my method has not got any impact on speed, even with
highly sophisticated models like "cms.gdml".


[FEATURES]

- Works with multiple constants in a single expression (tested)
- Should work with multiple *.gdml files (not tested)
- Works with constants with expressions (tested)

Example:

<constant name="size" value="100.0"/>
<constant name="width" value="2.0*size"/>
<constant name="height" value="4.0*size"/>
<constant name="depth" value="8.0*size"/>

*** Problem with rotation in boolean solids is fixed now ***

Originally, in GDML/Geant4, in a boolean composition of two solids only the second solid could be rotated.
This rotation has opposite winding compared to the rotation of the volume, so that the inverse of the rotation matrix is used.

Later the boolean solid was extended, using the "firstrotation" token, so that the first solid can be rotated too.
But in the code the matrix inversion was omitted for this rotation. By adding matrix inversion to this rotation
the problem is solved.


git-svn-id: https://root.cern.ch/svn/root/trunk@19562 27541ba8-7e3a-0410-8455-c3a389f83636