From Scott Snyder:
This is addressing some instances where root was showing O(N^2)
behavior with the number of defined classes/typedefs that were taking
a lot of time in the Atlas offline software.

Two instances of this were in TClass::Init and TCint::UpdateClassInfo.
Here, when a new class was created/updated, we would scan through
all existing classes looking for any where the names match after
full typedef expansion. This was particularly expensive since
the typedef expansion requires scanning through the entire
typedef table, potentially several times per class.

I addressed this by introducing a hash table, mapping from
the typedef-expanded name to the original class name.
An entry is made in the table only if those names
are different. So, to test for a duplicate, one can then
just look up the typedef-expanded name in the hash table,
rather than scanning all classes. For each hit,
one then looks up the class using the original name.
(We also need to do another lookup in the class list
using the typedef-expanded name if it's different.)
This eliminates the N^2 behavior here, and drastically
reduces the number of expensive typedef expansions.

One irregularity is that the two places that this happens
do the typedef expansion somewhat differently: one just expands
the name directly, but the other runs the name through
TClassEdit::ShortType with kDropStlDefault first. It's not
clear to me why this is done differently in the two places,
so i've retained this behavior, maintaining two separate
hash tables. It would be cheaper, though, if the same expansion
could be used in both places.

The other place that i made changes was in TCint::UpdateListOfTypes.
This was scanning through all the cint typedefs each time GetListOfTypes
was called to possibly add new ones to the root list. I changed this
so that it only scans the typedefs that are new since the last call,
going by the cint typedef index. But if the number of cint typedefs
has decreased since the last call, then we rescan the complete table.
This may not catch all possible cases where items have been deleted
from the cint table and then something else added --- to really
do that right, one would probably want a hook it the cint scrupto
routines.

This should now properly handle unload/reload in TCint::UpdateListOfTypes.
To do this, i added a counter in cint that gets bumped every time
the dictionary is rewound. We can thus test to see if that changed
to tell if a rewind has happened since the last call.

I also updated TClass so that the typedef mappings get removed
from the hash table when the TClass is deleted. To make that faster,
i used an additional status bit to flag if the TClass had created
any typedef map nodes.


git-svn-id: https://root.cern.ch/svn/root/trunk@19730 27541ba8-7e3a-0410-8455-c3a389f83636