NOTICE:
This version introduces support for transitioning classes 
from the mode where no class version is specified 
(and the checksum is always used for lookup) and a mode
where the user specifies an explicit version number for
the class (Allow for simplier, clearer data model evolution).

The class version can be specified in 3 different ways:

  a) Use rootcint and add a ClassDef macro in the class declaration:
	class MyClass {
           ...
           ClassDef(MyClass,10);
       };

  b) Use rootcint and add a RootClassVersion macro in the class source file:
	RootClassVersion(MyClass,10);

  c) Use genreflex and specify a ClassVersion property in the selection.xml :
       <class name="MyClass" ClassVersion="10"/>

Those 3 solutions set the class version of MyClass's TClass object 
to 10.

IMPORTANT:

In order to avoid a clash between this class version (and any future
class versions) and the class versions given by default to the 
non-versioned layout of the class, you MUST set this class version to
a number that is greater than the number of distinct non-versioned layout.
Otherwise you may see warning messages like:

Warning in <TStreamerInfo::BuildCheck>: 
   The class MyClass transitioned from not having a specified class version
   to having a specified class version (the current class version is 2).
   However too many different non-versioned layouts of the class have been
   loaded so far.  This prevent the proper reading of objects written with
   the class layout version 3, in particular from the file:
   myclass3.root.

Addition Details:

When loading a file containing a non-versioned layout of a class, this
layout is assigned the next 'free' number in the list of StreamerInfo.
In particular this means that if many files with non-versioned layout
of the class are loaded before the loading of a library containing a 
versioned class layout, the slot reserved for this version layout may 
already be occupied and you will get the following error message:

   The class MyClass transitioned from not having a specified class version
   to having a specified class version (the current class version is 2).
   However too many different non-versioned layouts of the class have
   already been loaded so far.  To work around this problem you can
   load fewer 'old' file in the same ROOT session or load the C++ library
   describing the class MyClass before opening the files or increase the version
   number of the class for example ClassDef(MyClass,3).
   Do not try to write objects with the current class definition,
   the files might not be readable.

Also if many files with non-versioned layout of the class are read
before a file with a versioned layout (and this number is too low), you
may get the following error message:

Warning in <TStreamerInfo::BuildCheck>: 
   The class MyClass transitioned from not having a specified class version
   to having a specified class version (the current class version is 2).
   However too many different non-versioned layouts of the class have been
   loaded so far.  This prevent the proper reading of objects written with
   the class layout version 3, in particular from the file:
   myclass3.root.

Additional note:
  For classes with no specified version (i.e. Foreign classes), the current
  TStreamerInfo is stored at index -1 in the list of StreamerInfo and the
  class version (gROOT->GetClass("MyClass")->GetClassVersion()) is set to -1.


git-svn-id: https://root.cern.ch/svn/root/trunk@21353 27541ba8-7e3a-0410-8455-c3a389f83636