From Paul: Change the behavior of G__get_methodhandle() when passed withConversion == 0.
The old behavior was to call G__get_ifunchandle_base() with a
match parameter of G__EXACT, and if that failed then try to
instantiate a function template, with exact match, and if that
failed then loop calling G__get_ifuncandle_base() with
parameters from G__EXACT up to and including G__STDCONV.  This
meant that we could match a function that did not exactly
match the passed prototype (and we called with G__EXACT
twice!).

The new behavior is to return in failure if the function
template instantiation attempt fails to find a match, the loop
trying the integral promotions and standard conversions is not
done.

So we must match the passed prototype exactly (unless the
template instantiation codes does something weird we do not
know about).

This fixes:

http://root.cern.ch/phpBB2/viewtopic.php?p=24563


git-svn-id: https://root.cern.ch/svn/root/trunk@21727 27541ba8-7e3a-0410-8455-c3a389f83636