Correct the schema evolution from a map<A,B> to a (non map, non multimap) collection of pair<A,B> (i.e vector<pair<A,B> >) when it is held by value in an object (as opposed to by pointer).
Conversion from vector<pair<A<B> > to map<A,B> is
now explicitly disabled (i.e. the vector will not be read).

Conversion for a map<A,B>* to a vector<pair<A,B> >* is
now explicitly disabled (i.e. the map will not be read).

TFile::MakeProject has been updated to generate the
needed dictionary for the pair.


git-svn-id: https://root.cern.ch/svn/root/trunk@21744 27541ba8-7e3a-0410-8455-c3a389f83636