From Hady Zalek: I've encountered several compilation and runtime errors when trying to port Reflex to Solaris and AIX:
AIX -- CC 5.8:
- Compilation fails under xlC with the following error: "DictionaryGenerator.cxx", line 42.16: 1540-0432 (S) The qualified declarator "ROOT::Reflex::operator<<" must refer to an existing declaration.
- Infinite loop on Reflex initialization due to incorrect map compare functions for 'string ' and 'const char '

Solaris -- xlC V8.0:
- Infinite loop on Reflex initialization due to incorrect map compare functions for 'string ' and 'const char '
- Segmentation fault due to incorrect buffer allocation with 'new char(buffer)' instead of 'new char[buffer]' in 'Tools.cxx':Demangle

Please find the first attached patch to correct the above issues...

The remaining patches are for a rewrite of Tools::NormalizeName to account for the mentioned platforms and make the demangling tests pass. I also took the freedom of adding an overloaded version in Tools.h that receives a 'const char *' parameter instead of a 'const std::string &'. The patch contains test cases for inclusion in test_ReflexBuilder_unit.cxx.


git-svn-id: https://root.cern.ch/svn/root/trunk@21765 27541ba8-7e3a-0410-8455-c3a389f83636