add the possibility to force the stub of a given function in the LinkDef
i.e.: #pragma link C++ options=stub function

This is actually implemented by turning off the function pointer in the second dictionary.
In that way, we wont get the symbol and the stub is generated.

Be careful, the way to signal this in the ifunc is a bit hacky (recheck to see if it can be
changed)


git-svn-id: https://root.cern.ch/svn/root/branches/dev@22148 27541ba8-7e3a-0410-8455-c3a389f83636