From Axel:
Fix G__class_autoload's test on a changed class name due to autoloading: hide 
the original (requetsed one) and do the name lookup. If it succeeds we have 
loaded a new representation of that class.
G__class_autoload now expects a int* ptagnum; it will be set to the tagnum of 
the loaded representation of the class (which might be different from the 
requested one)
Fixes autoload problems for STL.


git-svn-id: https://root.cern.ch/svn/root/branches/v5-18-00-patches@22164 27541ba8-7e3a-0410-8455-c3a389f83636