Fix a bug (re-introduced) found in roottest/root/io/multipleInherit/
When a method is found in multiple parents, we can't refer to it through just a simple index 
(the parents indices will have different values). So we flag it as ambiguous and do the full 
checking at run-time.

One possible way to do this is to look for the lef-most sibling index and propagate it through
the same inheritance level... (in that case... what kind of tree traversal do we do to find the 
rigth method? ... mmmm)


git-svn-id: https://root.cern.ch/svn/root/branches/dev@22662 27541ba8-7e3a-0410-8455-c3a389f83636