Tests in mathcore and mathmore changed so that now:
a) They return 0 if the test did go OK or another value otherwise. The return value in case of failure
   depends on the tests performed.

b) They tests that show some graphics or plots are changed now so that they will show the graphics by default, 
   unless the option -ng is passed as their only argument.

This two new functionalities are needed to make an script to automatically run all the tests.


git-svn-id: https://root.cern.ch/svn/root/branches/dev/mathDev@22880 27541ba8-7e3a-0410-8455-c3a389f83636