Insure that the TStreamerInfo a contained class that is to written memberwise is _not_ optimized.
Repair a few place where TStreamerInfo was wasting some space when the 
streamed collection was empty.  Increase TStreamerInfo version to 7 to 
distinguish the old and new cases.

Rename TCollectionProxyInfo::Environ::delete_temp to use_temp to clarify code/semantic
and fix its use to properly manage the life time of the temporary used (necessary in
the case of memberwise streaming (in non-split mode)).


git-svn-id: https://root.cern.ch/svn/root/trunk@22940 27541ba8-7e3a-0410-8455-c3a389f83636