fix certain synchronization issues...
the biggest problem is a circular dependancy TTreeCache<->TBranch that could cause a 
deadlock in certain (not yet found) cases. Fix it by invalidating the buffers instead of deleting them 
from the branch.

the thread checker is still showing certain warnings... but they seem to come from
different parts of root :/


git-svn-id: https://root.cern.ch/svn/root/branches/dev@22987 27541ba8-7e3a-0410-8455-c3a389f83636