For backward compatibility TStreamerInfo::BuildCheck compares the checksum of the on-file StreamerInfo not only to the current value of the class checksum but also to the checksum calculated using the older algorithms.
This patch extends this test to also be done when comparing 2 on-file StreamerInfo.

This is implemented via a new function TStreamerInfo::GetCheckSum(UInt_t) which
re-implement the TClass::GetCheckSum algorithms using the TStreamerElements 
instead of the RealData information.  [Note that due to slight different in 
class naming convention, the 2 methods are not completely interchangeable].

This removes spurrious warning message when loading 2 older files which
were written with 2 different version of the TClass CheckSum algorithm
(and the in-memory class's version is greater than both TStreamerInfos'
class version).


git-svn-id: https://root.cern.ch/svn/root/trunk@22996 27541ba8-7e3a-0410-8455-c3a389f83636