From Eddy: Three new methods have been implemented:
   Int_t   RescanLibraryMap();

Force a scan along the dynamic path again and load the *.rootmap files
in case the (DY)LD_LIBRARY_PATH has changed.

   Int_t   ReloadAllSharedLibraryMaps();

Reload for the loaded shared libraries their rootmap files.
This assumes that the library foo.[dll,a,so] has a rootmap named foo.rootmap.

   Int_t   UnloadAllSharedLibraryMaps();

Unload the rootmap entries for the loaded shared libraries.

I did not add anything to TSystem::SetDynamicPath() but in principle
the user should invoke after that operation RescanLibraryMap().


git-svn-id: https://root.cern.ch/svn/root/trunk@23039 27541ba8-7e3a-0410-8455-c3a389f83636