From Axel: there is a bug in GCC 4.3.0 that makes roottest/root/hist/operator/runsparse.C fail in the error calculation part. Funny enough it fails because TH3 messes up its errors (not THnSparse  :-) .
The problem stems from TH1::GetBinError():
   if (fSumw2.fN) return TMath::Sqrt(fSumw2.fArray[bin]);
does not get evaluated properly - it has _horribly_ wrong values, e.g.
  -5063: 0.42
  +5063: 0.648074
("-" is GCC 3.4.6, "+" is GCC 4.3.0)
after scaling a bin with a single entry by 0.42. So this is a real issue. It's possibly connected to the new math compiler intrinsics + pushing their results onto the stack.

Attached patch fixes this, by introducing a temporary.

      Double_t err2 = fSumw2.fArray[bin];
      return TMath::Sqrt(err2);

The patch also fixes THnSparse::Scale() - it was also scaling the entries.


git-svn-id: https://root.cern.ch/svn/root/trunk@23533 27541ba8-7e3a-0410-8455-c3a389f83636