New Implementation of the Genetic Algorithm in TMVA. It includes:
Changing the current implementation of GeneticPopulation and GeneticGenes so that, instead of using a std::multimap it can use std::vector and thus
make the current methods parallelizable.

Removed a certain amount of 'unused' code.

Parallelize the methods CalculateFitness, MakeChildren and Mutate with OMP.

Change the FitUtil file so that one of the tests becames thread safe.

To test the results, one can use: tmva/samples/TMVAGAExample and tmva/samples/TMVAGAExample2 from the TMVA
or the tests in math/genetic/tests/

Still to do:
  Solve the problem with the current implementation of GeneticFitter: It depends on old code that has to be 'reimplemented' or removed.

  Analyze the behaviour of the threads with Thread Profiler (doesn't work on Linux!!) or Shark in MacOSX


git-svn-id: https://root.cern.ch/svn/root/branches/dev@24196 27541ba8-7e3a-0410-8455-c3a389f83636