Correct the behavior of this code fragment:
   bool b = 0;
   bool* p = &b;
   bool d;
   d = *p;
   printf("*p: %d\n", *p);
   printf("d: %d\n", d);

The value of d was 1 instead of zero because the
code in variable assignment that was normalizing
the resultant value for a bool to 0 or 1 was testing
the actual value using the wrong bit width.

I have corrected a number of other places in the code
where G__BOOL4BYTE needs to be respected as well.

There was also a problem passing a bool as an argument
to a function parameter that is a reference to a bool
which has been corrected.

There is a new test bool01.cxx to verify correct
behavior in this case.

-- Paul Russo


git-svn-id: https://root.cern.ch/svn/root/branches/v5-18-00-patches@24612 27541ba8-7e3a-0410-8455-c3a389f83636