Fix improper calculation of interpreted function size.
The correct calculation is:

  end_definition_line - begin_definition_line + 1

What was being done was:

  end_definition_line - begin_declaration_line + 1

which is not correct, especially when the declaration
was first seen in a header file so that the line number
used is the line number in a different file!

Oddly enough this became visible through a bizzare
accident where the function size was calculated to
be -1, which is a magic value used to flag the
function as being compiled code instead of interpreted.
The result was the stub code jumping out into random
territory in an attempt to call a nonexistant
compiled function.

-- Paul Russo


git-svn-id: https://root.cern.ch/svn/root/trunk@24876 27541ba8-7e3a-0410-8455-c3a389f83636