ROOT special variables in cint5 can only be looked-up by prepending the variable name with '$'.  When they are created, they are entered into the appropriate G__struct entry's data member G__var_array with a hash code computed from the name with the '$' prepended, but the name entered in the G__var_array has the '$' removed, so the name and the hash code do not match.  This causes any search by name to fail, so these ROOT special variables are essentially hidden from normal lookup.  However they can still be found by looking for them with a '$' prepended to the name before hash code generation.
The problem for cint7 is that we do not use hash codes
for lookup so we cannot exactly emulate this behavior.

The solution this patch uses is to enter the name with the
'$' prepended, which is an illegal C++ variable name, so
it cannot be generated by normal user code.  This allows it
to be looked up, but still be essentially invisible to user
code.

-- Paul Russo


git-svn-id: https://root.cern.ch/svn/root/trunk@25056 27541ba8-7e3a-0410-8455-c3a389f83636