Thanks to Bertrand:
Now the fiteditor window is really fixed size. We couldn't fix the change of the size depending on the dimension of the object to be fitted though :( 
However, considering the overall size of the panel, I think it is good enough it we keep the size as it is ans the Y slider to appear and dissapear 
depending on the type of object.


git-svn-id: https://root.cern.ch/svn/root/branches/dev/mathDev@25076 27541ba8-7e3a-0410-8455-c3a389f83636