TEveGeoManagerHolder -------------------- Allow to also specify n-segments to be set to current geo-manager.
TEveGeoShape
------------
Add member 'Int_t fNSegments' to be passed to geo-manager when
painting the shape.
When it is set, enforce filling of raw section of buffer3d.

TGLScenePad
-----------
When raw section of buffer3d is filled for natively supported shapes
(spheres and tubes) prefer the passed tessellation over native shapes.


git-svn-id: https://root.cern.ch/svn/root/trunk@25477 27541ba8-7e3a-0410-8455-c3a389f83636