Fix tree->Draw("s1.value"); when the top level branch does not have a trailing dot (and hence the real branch name is only 'value').
TTreeFormula was properly parsing the string and finding the correct branch, however
it was recorded only the end branch name as the result of the parse (i.e. in the 
above case, just 'value') and hence UpdateFormulaLeaves was not restoring the leaf
correctly (and hence EvalInstance returned 0s).
To solve the issue we now record the 'full pathname' of the branch (i.e. including
the omitted top level branch names) and we upgrade TTree::FindBranch and
TBranchElement::FindBranch to be able to handle those path.

Fix TTree::Scan("*") to properly requested the distinct leaf even if the branch
name are ambiguous (previously it would repetively print the first match).


git-svn-id: https://root.cern.ch/svn/root/trunk@25572 27541ba8-7e3a-0410-8455-c3a389f83636