Now once the fit is produced inside the FitPanel, the sliders for the ranges in the three dimensions will stay the same as before the fitting.
There are two solucions for this problem, one is disconnecting the slot for the RangeAxisChanged of the fParentPad, 
that would call UpdateGUI and this one would reset the status of the sliders to the original one. The second solution 
is the one applied in this patch, that stores the previous values of the sliders and then restablish them after the 
update of the canvas.

The first one is more elegant and is the one that should be used, however, the disconnection of the slot was already 
done and I tried to disconnect it in several places of the code (i.e. specially right beofore the update of the 
canvas) and it didn't work. The reason is unknown to me :S


git-svn-id: https://root.cern.ch/svn/root/branches/dev@25617 27541ba8-7e3a-0410-8455-c3a389f83636