When a new function is selected from the TGComboBox list with the available predefined functions, the parameters of the functions are reseted only if both functions, the old and the new ones, differ in their number.
Before this modification if the function was fitted with a sqroot and accidentally I reselect sqroot from the list, 
all the parameters would be lost. Now they are kept. If the user wants to still reset them, they will have to do it 
manually form the parameters dialog of the FitPanel.


git-svn-id: https://root.cern.ch/svn/root/branches/dev@25621 27541ba8-7e3a-0410-8455-c3a389f83636