Merge revs 25435:25641 from the fireworks branch.
Many improvements in text and axis rendering for orthographic views,
for axes in non-linear projections and for the dedicated lego view.

TGLSceneBase
------------
Add data-member 'Bool_t fSelectable' with get/set methods.
If this is set to false, the scene is not rendered during selection
thus preventing any of its elements to be selected.

TGLViewerBase
-------------
In PreRender() honour the selectable state of a scene when selecting
which scenes to include in rendering.

TGLSAViewer
-----------
Remove extra border introduced by usage of kSunkenFrame.

TGLEmbeddedViewer
-----------------
Add optional argument 'Int_t border' to the constructor.


git-svn-id: https://root.cern.ch/svn/root/trunk@25642 27541ba8-7e3a-0410-8455-c3a389f83636