The function CheckRange has been removed. It seemed to check the range of the function in the X axis with respect to the sliders shown in the fit panel. This funcionality has been moved into the UpdateGUI function were the rest of the axis were being checked.
I have done several tests changing the range of the axis of the plot
and it seems like the fit panel still works as before. However, more
tests would be convinient.

Next I will try to remove some member variables that were used inside
some of the removed methods and make the code a bit simpler.


git-svn-id: https://root.cern.ch/svn/root/branches/dev@25653 27541ba8-7e3a-0410-8455-c3a389f83636