The update and set up of the sliders has been unified into the UpdateGUI method, simplifying the DoReset one. This also allowed the reduction of some data members of the class.
I also changed a check for a histogram obtained from the object to be
fitted. It look like if the comparision was not properly handled and
it would crash or produce incorrect behaviour in such a case. I added
an assertion. If everything is fine it should be no problem,
otherwise, we'll know inmediatly where the problem is.


git-svn-id: https://root.cern.ch/svn/root/branches/dev/mathDev@25662 27541ba8-7e3a-0410-8455-c3a389f83636