import from the branch the lates fixed from David:
The update and set up of the sliders has been unified into the
UpdateGUI method, simplifying the DoReset one. This also allowed the
reduction of some data members of the class.

The function CheckRange has been removed. 
This funcionality has been moved into the UpdateGUI
function were the rest of the axis were being checked.

The Parameters rxmin and rxmax where removed from the TFitParamertersDialog, 
as they were not used nor needed.


git-svn-id: https://root.cern.ch/svn/root/trunk@25740 27541ba8-7e3a-0410-8455-c3a389f83636