Just as a CheckPoint!
First steps in the implementation of a TGComboBox for selecting the object to be fitted. Now the combo box is created but still nothing is properly done.
The fit panel outfit has been slightly changed but some more changed will come in the next submits.

There is also a couple of changes in the previous code to make it more legible.

This commit still contains the debug prints.


git-svn-id: https://root.cern.ch/svn/root/branches/dev/mathDev@25790 27541ba8-7e3a-0410-8455-c3a389f83636