This patch corrects bug #42983 reported in Savannah.
Release Notes:

The G__exec_binopr macro in expr.cxx stacks a multiply by
negative one on the expression stack when it sees an unary minus
during the evaluation of the power operator ("**" and "@").
We raise the precedence of this stacked multiply to be the
same as the power operator so that it will be evaluated during
the power operation and not afterwards.


git-svn-id: https://root.cern.ch/svn/root/trunk@25871 27541ba8-7e3a-0410-8455-c3a389f83636