Termporary changes made to TGComboBox.cxx and TGComboBox.h
There were two problems that I detected for the use of this class in the fitpanel prototype:

1 The TGComboBox did not find an entry if this was currently selected by the user.

2 The size of the listbox was not properly set after an insertion.

Bertrand solved the first problem and he is now checking the validity of the changed I made for the second. Still its his decision how the final
code will look like, so when the mathDev branch is merged with the trunk, the lattest shold remain!


git-svn-id: https://root.cern.ch/svn/root/branches/dev@26092 27541ba8-7e3a-0410-8455-c3a389f83636